+++ game-frontend/app/Buildings/types/Storehouse.js
@@ -2,8 +2,8 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 const Building_1 = require("../Building");
 class Storehouse extends Building_1.default {
-    constructor(position) {
-        super(position, 'storehouse');
+    constructor(position, id) {
+        super(id, position, 'storehouse');
     }
 }
 Storehouse.asset = 'storehouse.babylon';