+++ game-frontend/app/Commands/BuildingUpdateCommand.js
@@ -12,9 +12,8 @@ class BuildingUpdateCommand extends Command_1.default {
         return 'building.update';
     }
     execute(req) {
-        console.log(req._id);
         if (!exports.builtBuildings[req._id]) {
-            exports.builtBuildings[req._id] = new buildingMapping_1.default[req.type](req.position);
+            exports.builtBuildings[req._id] = new buildingMapping_1.default[req.type](req.position, req._id);
         }
         exports.builtBuildings[req._id].setHealt(req.currentHealth, req.maxHealth);
     }