+++ game-frontend/app/main.js
@@ -25,7 +25,7 @@ function createWindow() {
         slashes: true,
     }));
     // Open the DevTools.
-    //mainWindow.webContents.openDevTools();
+    mainWindow.webContents.openDevTools();
     // Emitted when the window is closed.
     mainWindow.on("closed", () => {
         // Dereference the window object, usually you would store windows