+++ game-frontend/app/ui/actions/CreateCharacter.js
@@ -1,6 +1,7 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 const connection_1 = require("../../services/connection");
+const panel_1 = require("../panel/panel");
 class CreateCharacter {
     constructor(element) {
         const type = element.dataset.type;
@@ -9,6 +10,7 @@ class CreateCharacter {
         }
         connection_1.default.socket.emit('character.create', {
             type: type,
+            buildingId: panel_1.default.selectedBuildingId,
         });
     }
 }