+++ game-frontend/app/ui/panel/panel.js
@@ -22,12 +22,17 @@ class Panel {
      */
     onClick() {
         var pickResult = Game_1.default.gameScene.scene.pick(Game_1.default.gameScene.scene.pointerX, Game_1.default.gameScene.scene.pointerY);
+        this.selectedBuildingIsReady = true;
         if (!pickResult.pickedMesh || !pickResult.pickedMesh.metadata || !pickResult.pickedMesh.metadata.key) {
             return;
         }
         let key = pickResult.pickedMesh.metadata.key;
         if (!pickResult.pickedMesh.metadata.isBuilding) {
             key = 'default';
+        }
+        else {
+            this.selectedBuildingId = pickResult.pickedMesh.metadata.buildingId;
+            this.selectedBuildingIsReady = pickResult.pickedMesh.metadata.isReady;
         }
         this.hideAllPanels();
         this.showPanel(key);
@@ -37,13 +42,30 @@ class Panel {
         this.showPanel('default');
     }
     showPanel(key) {
-        document.querySelector(`[data-panel-key="${key}"]`).classList.remove('is-hidden');
+        const element = document.querySelector(`[data-panel-key="${key}"]`);
+        element.classList.remove('is-hidden');
+        if (false === this.selectedBuildingIsReady) {
+            element.classList.add('is-not-ready');
+        }
     }
     hideAllPanels() {
         document.querySelectorAll('[data-panel-key]')
             .forEach((element) => {
             element.classList.add('is-hidden');
+            element.classList.remove('is-not-ready');
         });
+    }
+    get selectedBuildingId() {
+        return this._selectedBuildingId;
+    }
+    set selectedBuildingId(selectedBuildingId) {
+        this._selectedBuildingId = selectedBuildingId;
+    }
+    get selectedBuildingIsReady() {
+        return this._selectedBuildingIsReady;
+    }
+    set selectedBuildingIsReady(value) {
+        this._selectedBuildingIsReady = value;
     }
 }
 const panel = Panel.getInstance();