+++ game-frontend/CHANGES.md
@@ -1,5 +1,10 @@
 # CHANGELOG for Time of Changes
 
+#### 0.1.1 (2018-10-19)
+ - BUGFIX: Fixed multiplayer bug.
+ - FEATURE: Added start characters.
+ - FEATURE: Characters can be created only on built buildings.
+
 #### 0.1.0 (2018-10-19)
  - BUGFIX: Fixed a multiplayer bug that hides buildings from other peoples.
  - FEATURE: Added a new panel interface.